Fixes according review
