+++ src/config/settings.rs
@@ -1,8 +1,10 @@
+use crate::PonyError;
+use crate::Result;
 use default_net::{get_default_interface, get_interfaces};
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
+use serde::Serializer;
 use std::env;
-use std::error::Error;
 use std::fs;
 use std::net::Ipv4Addr;
 use uuid::Uuid;
@@ -181,8 +183,18 @@ pub struct LoggingConfig {
     pub level: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Default)]
+#[derive(Clone, Debug, Deserialize)]
 pub struct NodeConfig {
+    pub env: String,
+    pub hostname: String,
+    pub default_interface: String,
+    pub address: Ipv4Addr,
+    pub uuid: Uuid,
+    pub label: String,
+}
+
+#[derive(Clone, Debug, Deserialize, Default)]
+pub struct NodeConfigRaw {
     #[serde(default = "default_env")]
     pub env: String,
     pub hostname: Option<String>,
@@ -195,50 +207,44 @@ pub struct NodeConfig {
 }
 
 impl NodeConfig {
-    // FIXME(qezz): validate() mutates self, while in theory it should not.
-    // Common solution is to have 2 types for config, e.g. `RawNodeConfig`
-    // and `NodeConfig`.
-    //
-    // Then: NodeConfig::from_raw(raw: RawNodeConfig) -> NodeConfig { ... }
-    pub fn validate(&mut self) -> Result<(), Box<dyn Error>> {
-        if self.hostname.is_none() {
+    pub fn from_raw(raw: NodeConfigRaw) -> Result<NodeConfig> {
+        let hostname = if raw.hostname.is_none() {
             match env::var("HOSTNAME") {
-                Ok(hostname) => {
-                    self.hostname = Some(hostname);
-                }
+                Ok(hostname) => hostname,
                 Err(_) => {
-                    return Err("Validation error: missing hostname (set $HOSTNAME env or specify in config)".into());
+                    return Err(PonyError::Custom("Validation error: missing hostname (set $HOSTNAME env or specify in config)".into()));
                 }
             }
-        }
+        } else {
+            raw.hostname.unwrap()
+        };
 
-        if let Some(ref interface_name) = self.default_interface {
+        let (address, interface) = if let Some(ref interface_name) = raw.default_interface {
             let interfaces = get_interfaces();
             if let Some(interface) = interfaces.iter().find(|i| &i.name == interface_name) {
                 match interface.ipv4.first() {
-                    Some(network) => self.address = Some(network.addr),
+                    Some(network) => (network.addr, interface_name.to_string()),
                     None => {
-                        return Err(
+                        return Err(PonyError::Custom(
                             "Validation error: Cannot get IPv4 address for the specified interface"
                                 .into(),
-                        )
+                        ));
                     }
                 }
             } else {
-                return Err(
+                return Err(PonyError::Custom(
                     format!("Validation error: Interface {} not found", interface_name).into(),
-                );
+                ));
             }
         } else {
             match get_default_interface() {
                 Ok(interface) => {
-                    self.default_interface = Some(interface.name);
                     if interface.ipv4.is_empty() {
-                        return Err(
+                        return Err(PonyError::Custom(
                             "Validation error: Cannot get IPv4 address of default interface".into(),
-                        );
+                        ));
                     } else {
-                        self.address = Some(interface.ipv4[0].addr);
+                        (interface.ipv4[0].addr, interface.name)
                     }
                 }
                 Err(e) => {
@@ -247,9 +253,16 @@ impl NodeConfig {
                     )
                 }
             }
-        }
+        };
 
-        Ok(())
+        Ok(NodeConfig {
+            env: raw.env,
+            hostname: hostname,
+            default_interface: interface,
+            address: address,
+            uuid: raw.uuid,
+            label: raw.label,
+        })
     }
 }
 
@@ -297,9 +310,11 @@ pub struct ZmqSubscriberConfig {
 }
 
 impl ZmqSubscriberConfig {
-    pub fn validate(self) -> Result<(), Box<dyn Error>> {
+    pub fn validate(self) -> Result<()> {
         if !self.endpoint.starts_with("tcp://") {
-            return Err("ZMQ endpoint should start with tcp://".into());
+            return Err(PonyError::Custom(
+                "ZMQ endpoint should start with tcp://".into(),
+            ));
         }
         Ok(())
     }
@@ -312,16 +327,18 @@ pub struct ZmqPublisherConfig {
 }
 
 impl ZmqPublisherConfig {
-    pub fn validate(self) -> Result<(), Box<dyn Error>> {
+    pub fn validate(self) -> Result<()> {
         if !self.endpoint.starts_with("tcp://") {
-            return Err("ZMQ endpoint should start with tcp://".into());
+            return Err(PonyError::Custom(
+                "ZMQ endpoint should start with tcp://".into(),
+            ));
         }
         Ok(())
     }
 }
 
 pub trait Settings: Sized {
-    fn read_config<T: DeserializeOwned>(config_file: &str) -> Result<T, Box<dyn Error>> {
+    fn read_config<T: DeserializeOwned>(config_file: &str) -> Result<T> {
         let config_str = fs::read_to_string(config_file)?;
         let settings: T = toml::from_str(&config_str)?;
         Ok(settings)
@@ -337,7 +354,7 @@ pub trait Settings: Sized {
         }
     }
 
-    fn validate(&mut self) -> Result<(), String>;
+    fn validate(&self) -> Result<()>;
 }
 
 #[derive(Clone, Debug, Deserialize, Default)]
@@ -347,7 +364,7 @@ pub struct ApiSettings {
     #[serde(default)]
     pub debug: DebugConfig,
     #[serde(default)]
-    pub node: NodeConfig,
+    pub node: NodeConfigRaw,
     #[serde(default)]
     pub logging: LoggingConfig,
     #[serde(default)]
@@ -360,22 +377,7 @@ pub struct ApiSettings {
     pub carbon: CarbonConfig,
 }
 
-impl Settings for ApiSettings {
-    fn validate(&mut self) -> Result<(), std::string::String> {
-        self.node
-            .validate()
-            .map_err(|e| format!("Node configuration validation error: {}", e))?;
-
-        self.zmq
-            .clone()
-            .validate()
-            .map_err(|e| format!("Zmq configuration validation error: {}", e))?;
-
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Default)]
+#[derive(Clone, Debug, Deserialize)]
 pub struct AgentSettings {
     #[serde(default)]
     pub debug: DebugConfig,
@@ -392,22 +394,21 @@ pub struct AgentSettings {
     #[serde(default)]
     pub zmq: ZmqSubscriberConfig,
     #[serde(default)]
-    pub node: NodeConfig,
+    pub node: NodeConfigRaw,
     #[serde(default)]
     pub api: ApiAccessConfig,
 }
 
 impl Settings for AgentSettings {
-    fn validate(&mut self) -> Result<(), String> {
-        self.node
-            .validate()
-            .map_err(|e| format!("Node configuration validation error: {}", e))?;
+    fn validate(&self) -> Result<()> {
+        self.zmq.clone().validate()?;
+        Ok(())
+    }
+}
 
-        self.zmq
-            .clone()
-            .validate()
-            .map_err(|e| format!("Zmq configuration validation error: {}", e))?;
-
+impl Settings for ApiSettings {
+    fn validate(&self) -> Result<()> {
+        self.zmq.clone().validate()?;
         Ok(())
     }
 }
